+++ addons/dms_access_fix/__init__.py
@@ -1,3 +1,4 @@
 # -*- coding: utf-8 -*-
 from . import models
+from . import controllers
 from .hooks.post_init import post_init